vaudtax/vaudtax_backend/requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.14.0
django-cors-headers>=4.3.0
djangorestframework-simplejwt>=5.3.0
psycopg2-binary>=2.9.9
django-environ>=0.11.2
Pillow>=10.2.0
openai>=1.12.0
zod>=3.22.4
requests>=2.31.0
